Take default value for Enum from allowed values
If no default is explicitly specified and allow_none is False (the
default), Enum currently causes an error if you instantiate the object
owning it without supplying a value for it.

This changes makes it take the first of the possible values as the
default instead. If you want to force the code instantiating the class
to supply a value, you can create the Enum with `default_value=None`.