Ignore unspecified default value until first access
Alternative to gh-13

Previously, if:

- The traitlet class does not specify a default value,
- The use of the traitlet does not provide a default value,
- The class using the traitlet does not have a _foo_default method, and
- The instantiation of that class does not provide a value for the
traitlet

then the traitlet would try to set its value to the undefined default,
usually causing an error. With this change, it will skip that, so the
traitlet will have no value. If it is accessed before a value has been
set, the error will occur then instead.

This allows Enums with no default value, which I want for kernelspec.

`Instance` traitlets are not affected - they will still create an
instance of the specified class when the class using them is
instantiated.