Don't create default values for Instance until required
This is an alternative to #23 and #28

There is already a mechanism for dynamic default values with
_foo_default methods, which are not called unless needed. This makes the
default values for Instance traitlets use that same mechanism, avoiding
instantiating classes unless those traitlets are explicitly requested
without being assigned first.

Along the way, two other changes to behaviour, and hence to tests, came
in. Dict traitlets are based on Instance, so their behaviour changes in
the same way. And a Type() traitlet with no class specified now gets a
default value of object instead of None, because the former is valid
when allow_none is True.

It was also possible to clean up some parts of the machinery.