avoid infinite recursion cross-validating defaults
When validating defaults, load initial values prior to cross-validation

Downside: this means that accessing other values during validation
can return non-final values in the event of cross-validation cycles
(`@validate(a)` depends on self.b and `@validate(b)` depends on self.a).
But this is only true for validated default values with cross-references
where validation would mutate the default values,
which is the same as having invalid default values.