Expose loaded config files, make load idempotent
This change enables applications to implement periodic reloads of
updated configuration files without requiring restarts.  This is
intended for long-running, service-oriented, applications that
cannot afford to be shutdown yet may need to have their configurations
updated.

This change exposes the _loaded_config_files list on the application
class instance via a property.  This was done so as to not allow
side-affects to creep in from clients.

More importantly, is the prevention of duplicate config file entries
that can occur when load_config_file is called multiple times. This i
enables the method's idempotency - otherwise the list of loaded config
files increased on each (periodic) call.