Backport applicable portion of PR 341
This change only takes a very small port of PR 341.  Specifically,
its a small portion of commit eaf1ae8 that introduces the List
`_loaded_config_files` and returns a tuple from `load_config_file()`.

This was necessary since the rest of this commit and the other 5
commits introduced API-level changes that were dependent on other
changes not pertinent to a minor release.