Allow compounding of Lazy traitlets across configuration files.
This is mostly intesting in the case of systemwide configuration in
/etc/local... to provide default configuration to user and still allow
them to extend their configuration.

Current traitlets behavior is as follow.

   /etc/local/config.py ::  c.Foo.value = ['a']
   ~/.jupyter/config.py ::  c.Foo.value.append('b')

Final seen value will be `['b']`

With the changes included here, you would get `['a', 'b']`

It also become possible to do:

   /etc/local/config.py ::  c.Foo.value.append('a')
   ~/.jupyter/config.py ::  c.Foo.value.append('b')

In which case the end result will be `['default', 'a', 'b']`.

This feel like a more reasonable (and requested) behavior than the
current.