Ensure default Set values are sorted for a reproducible build
Whilst working on the Reproducible Builds effort [0] we noticed
that traitlets generates non-reproducible output which is affecting
the reproducibility of other packages. For example, in nbconvert:

  -<dd><p class="first">Default: {'image/jpeg',</span> <span class="pre">'image/svg+xml',</span> [..]
  +<dd><p class="first">Default: {'image/svg+xml',</span> <span class="pre">'application/pdf',</span> [..]

(From https://tests.reproducible-builds.org/debian/rb-pkg/unstable/
amd64/nbconvert.html on 20191014)

This is due to it not iterating over a Set traitlet type in a
deterministic ordering when generating the "Default:" human-readable
string.

This patch ensures that we display the Set traits defaults in a
sorted, and thus reproducible, manner.

  [0] https://reproducible-builds.org/