Update URL for the Traits library
The Traitlets README refers to the Traits library and gives a `code.enthought.com` URL, but the resources at `code.enthought.com` are unmaintained and out of date. This PR updates that URL to something more current. This URL _should_ be reasonably stable going forward. (Another alternative that ought to be fairly safe stability-wise would be to refer to the PyPI page: https://pypi.org/project/traits/ .)