Initial traitlets subcommand support for argcomplete
argcomplete's strategy is to call the python script with
no arguments e.g. len(sys.argv) == 1, run until the ArgumentParser
is constructed and determine what completions are available.
On the other hand, traitlet's subcommand-handling strategy
is to check sys.argv[1] and see if it matches a subcommand,
and if so then dynamically load the subcommand app and initialize
it with sys.argv[1:].

Write a couple of helper functions to reconcile this by:

1. retrieving tokens from $COMP_LINES, etc, and setting it to argv
2. if traitlets descends into a subcommand, increment index passed
   via env var to argcomplete to mark where command starts

There's quite a few caveats to this approach. For example, it only
is evaluated currently when `App.initialize()` is passed with
`argv=None` (the default). If `argv` is explicitly passed, then
the `argcomplete`-specific handling is skipped currently.

More details in:
https://github.com/ipython/traitlets/pull/811#issuecomment-1345535450

Some additional minor cleanup with respect to subcommands typing,
examples, and documentation.