Maint: Remove potentially extra usage of Unicode.
As far as I can tell setting this to `Unicode` does not achieve much.
For example the IPython application that derives from it set it to a
plain string and it works.

Same for description.

From my dev folder I haven't found any usage of name or description that
needs this to be a `Unicode`, we we might as well make it always a str,
which simplify a bit the codebase/types.

Usually the simpler the better.

It is _technically_ backward incompatible.

See https://github.com/ipython/ipython/pull/13895 that remove the usage
of Unicode in description